Renamed example test files so they dont get run with other  maven invocations
