+++ whois-commons/src/test/java/net/ripe/db/whois/common/profiles/usingspring/ripe/WhoisVariantRipeIntegrationUST.java
@@ -1,4 +1,4 @@
-package net.ripe.db.whois.common.profiles.usingspring.apnic;
+package net.ripe.db.whois.common.profiles.usingspring.ripe;
 
 import net.ripe.db.whois.common.IntegrationTest;
 import net.ripe.db.whois.common.profiles.usingspring.ModifiedAbstractDatabaseHelperTest;
@@ -21,10 +21,11 @@ import static org.junit.Assert.assertTrue;
 
 @ContextConfiguration(locations = {"classpath:applicationContext-commons-test-usingprofiles.xml"})
 @Category(IntegrationTest.class)
-public class WhoisVariantApnicIntegrationTest extends ModifiedAbstractDatabaseHelperTest {
-    private static final Logger LOGGER = LoggerFactory.getLogger(WhoisVariantApnicIntegrationTest.class);
+public class WhoisVariantRipeIntegrationUST extends ModifiedAbstractDatabaseHelperTest {
+    private static final Logger LOGGER = LoggerFactory.getLogger(WhoisVariantRipeIntegrationUST.class);
 
-    @Rule public ExpectedException exception = ExpectedException.none();
+    @Rule
+    public ExpectedException exception = ExpectedException.none();
 
     @BeforeClass
     public static void beforeClass() throws Exception {
@@ -32,16 +33,16 @@ public class WhoisVariantApnicIntegrationTest extends ModifiedAbstractDatabaseHe
     }
 
     @Test
-    public void load_apnic_bean_success() {
-        WhoisVariantBeanApnic whoisVariantBeanApnic = (WhoisVariantBeanApnic)applicationContext.getBean(StringUtils.uncapitalize(WhoisVariantBeanApnic.class.getSimpleName()));
-        assertTrue(whoisVariantBeanApnic.isApnic());
-        assertNotNull(whoisVariantBeanApnic.getUpdateDao());
-    }
-    @Test
-    public void load_ripe_bean_fail() {
+    public void load_apnic_bean_fail() {
         exception.expect(NoSuchBeanDefinitionException.class);
-        WhoisVariantBeanRipe whoisVariantBeanRipe = (WhoisVariantBeanRipe)applicationContext.getBean(StringUtils.uncapitalize(WhoisVariantBeanRipe.class.getSimpleName()));
+        WhoisVariantBeanApnic whoisVariantBeanApnic = (WhoisVariantBeanApnic)applicationContext.getBean(StringUtils.uncapitalize(WhoisVariantBeanApnic.class.getSimpleName()));
     }
 
+    @Test
+    public void load_ripe_bean_success() {
+        WhoisVariantBeanRipe whoisVariantBeanRipe = (WhoisVariantBeanRipe)applicationContext.getBean(StringUtils.uncapitalize(WhoisVariantBeanRipe.class.getSimpleName()));
+        assertTrue(whoisVariantBeanRipe.isRipe());
+        assertNotNull(whoisVariantBeanRipe.getUpdateDao());
+    }
 
 }
